Use double instead of float.

+++ Jig/Matrix.h
@@ -7,14 +7,14 @@ namespace Jig
 	{
 	public:
 		void MakeIdentity();
-		void MakeRotationX(float angle);
-		void MakeRotationY(float angle);
-		void MakeRotationZ(float angle);
-		void MakeRotationYXZ(float ry, float rx, float rz);
-		void MakeScale(float sx, float sy, float sz);
+		void MakeRotationX(double angle);
+		void MakeRotationY(double angle);
+		void MakeRotationZ(double angle);
+		void MakeRotationYXZ(double ry, double rx, double rz);
+		void MakeScale(double sx, double sy, double sz);
 		//void MakeZAxisToVector( const Vector& vec );
 		void MakeInverse(const Matrix& src);
-		//	void MakeRotationVector(float angle, const Vector& axis);
+		//	void MakeRotationVector(double angle, const Vector& axis);
 
 		void MultRight(const Matrix& a);
 		void MultLeft(const Matrix& a);
@@ -24,20 +24,20 @@ namespace Jig
 		void MultPointInverse(Vec3& dst) const;
 		void MultPointInverse(Vec3& dst, const Vec3& src) const;
 
-		float* GetData() { return mat; }
-		const float* GetData() const { return mat; }
+		double* GetData() { return mat; }
+		const double* GetData() const { return mat; }
 
 		//	void GetZAxis( Vector& zVec ) const; 
 
 		void Print() const;
-		void Get4x4(float m[16]) const;
+		void Get4x4(double m[16]) const;
 		void glMultMatrix() const;
 
 	private:
 		static void Mult(Matrix* pDst, const Matrix& a, const Matrix& b);
-		float mat[9];
+		double mat[9];
 
-		float Clip(float a);
+		double Clip(double a);
 	};
 
 	class MatrixGL
